requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.2
pandas>=2.2.0
numpy>=1.26.0
httpx>=0.26.0
python-dotenv>=1.0.0
